sdk: python: Move examples into sdk’s path
