--- examples/python/client-simple/README.md
@@ -1,8 +1,0 @@
-# How to use this example
-
-```sh
-python3 -m venv ../../../.venv
-source ../../../.venv/bin/activate
-python3 -m pip install -r requirements.txt
-python3 say.py "Simple is better than complex"
-```
