--- examples/python/client-simple/requirements.txt
@@ -1,1 +1,0 @@
--e ../../../sdk/python
