+++ sdk/python/examples/client-simple/say.py
@@ -1,6 +1,7 @@
 import sys
 
 import anyio
+
 import dagger
 
 
@@ -25,4 +26,7 @@ async def main(args: list[str]):
 
 
 if __name__ == "__main__":
+    if not sys.argv[1:]:
+        print("What do you want to say?", file=sys.stderr)
+        sys.exit(1)
     anyio.run(main, sys.argv[1:])
