+++ sdk/python/experimental/hello/requirements.txt
@@ -1,0 +1,1 @@
+dagger-io
